Count time convert point

+++ core/src/commonMain/kotlin/com/noljanolja/core/exchange/domain/domain/ExchangeRate.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ExchangeRate(
     val coinToPointRate: Long = 0L,
-    val rewardRecurringAmount: Long = 0L,
+    val rewardRecurringAmount: Int = 0,
 )
